Added goodRx client. Don't have developer key but am using the searches they use on their ui
